requirements.txt
aiohttp>=3.9.5
aiosqlite>=0.20.0
python-dotenv>=1.0.1
discord.py>=2.4.0
openai>=1.30.0
tzdata>=2024.1
